% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpara_adjVUS.R
\name{vus_np1}
\alias{vus_np1}
\title{First nonparametric estimation method of VUS}
\usage{
vus_np1(mu, sig, errors, ...)
}
\arguments{
\item{mu}{a vector of three estimated conditional means for given covariate's value.}

\item{sig}{a vector of three estimated conditional variances for given covariate's value.}

\item{errors}{a list of three estimated errors in regression models.}

\item{...}{additional arguments supplied to the \code{npudist()} and \code{npudens()} functions.}
}
\description{
\code{vus_np1} computes covariate-specific estimates of the VUS by using integrate with approximated distribution.
}
