% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpara_adjVUS.R
\name{vus_np2}
\alias{vus_np2}
\title{Second nonparametric estimation method of VUS}
\usage{
vus_np2(mu, sig, errors)
}
\arguments{
\item{mu}{a vector of three estimated conditional means for given covariate's value.}

\item{sig}{a vector of three estimated conditional variances for given covariate's value.}

\item{errors}{a list of three estimated errors in regression models.}
}
\description{
\code{vus_np2} computes covariate-specific estimates of the VUS by using working samples.
}
